[refactor] Relocate error layout
- error layout 파일을 layouts 폴더로 이동
